requirements.txt
matplotlib==3.7.1
numpy==1.24.3
Pillow==9.4.0
torch==1.12.0
torchvision==0.13.0
tqdm==4.65.0
